+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master (unreleased)
 	
 ### New Features 
+
+    - Add a bookmarking system
             
 ### Changes